+++ tests/Feature/compiler-test-source/absolute-or-relative.cpp
@@ -1,0 +1,11 @@
+#include <stdio.h>
+
+#include "/etc/hosts"
+
+#include "../some/totally/ok/include.h"
+
+int main()
+{
+    printf("Hello,World\n");
+    return 0;
+}